+++ torchrl/data/__init__.py
@@ -46,3 +46,4 @@ from .tensor_specs import (
     UnboundedContinuousTensorSpec,
     UnboundedDiscreteTensorSpec,
 )
+from .utils import check_no_exclusive_keys, consolidate_spec
